BV-9424

+++ code/modules/jobs/job_types/_job.dm
@@ -129,36 +129,37 @@
 			else
 				M.client.prefs.equipped_gear -= gear
 
-	if(gear_leftovers.len)
-		for(var/datum/gear/G in gear_leftovers)
-			var/metadata = M.client.prefs.equipped_gear[G.display_name]
-			var/item = G.spawn_item(null, metadata)
-			var/atom/placed_in = human.equip_or_collect(item)
+	if(gear_leftovers)
+		if(gear_leftovers.len)
+			for(var/datum/gear/G in gear_leftovers)
+				var/metadata = M.client.prefs.equipped_gear[G.display_name]
+				var/item = G.spawn_item(null, metadata)
+				var/atom/placed_in = human.equip_or_collect(item)
 
-			if(istype(placed_in))
-				if(isturf(placed_in))
-					to_chat(M, "<span class='notice'>Placing [G.display_name] on [placed_in]!</span>")
-				else
-					to_chat(M, "<span class='noticed'>Placing [G.display_name] in [placed_in.name]]")
-				continue
+				if(istype(placed_in))
+					if(isturf(placed_in))
+						to_chat(M, "<span class='notice'>Placing [G.display_name] on [placed_in]!</span>")
+					else
+						to_chat(M, "<span class='noticed'>Placing [G.display_name] in [placed_in.name]]")
+					continue
 
-			if(H.equip_to_appropriate_slot(item))
-				to_chat(M, "<span class='notice'>Placing [G.display_name] in your inventory!</span>")
-				continue
-			if(H.put_in_hands(item))
-				to_chat(M, "<span class='notice'>Placing [G.display_name] in your hands!</span>")
-				continue
+				if(H.equip_to_appropriate_slot(item))
+					to_chat(M, "<span class='notice'>Placing [G.display_name] in your inventory!</span>")
+					continue
+				if(H.put_in_hands(item))
+					to_chat(M, "<span class='notice'>Placing [G.display_name] in your hands!</span>")
+					continue
 
-			var/obj/item/storage/B = (locate() in H)
-			if(B)
-				G.spawn_item(B, metadata)
-				to_chat(M, "<span class='notice'>Placing [G.display_name] in [B.name]!</span>")
-				continue
+				var/obj/item/storage/B = (locate() in H)
+				if(B)
+					G.spawn_item(B, metadata)
+					to_chat(M, "<span class='notice'>Placing [G.display_name] in [B.name]!</span>")
+					continue
 
-			to_chat(M, "<span class='danger'>Failed to locate a storage object on your mob, either you spawned with no hands free and no backpack or this is a bug.</span>")
-			qdel(item)
+				to_chat(M, "<span class='danger'>Failed to locate a storage object on your mob, either you spawned with no hands free and no backpack or this is a bug.</span>")
+				qdel(item)
 
-		qdel(gear_leftovers)
+			qdel(gear_leftovers)
 
 /datum/job/proc/announce(mob/living/carbon/human/H)
 	if(head_announce)
